apps/rag/requirements.txt
langchain
langchain_community
langchainhub
openai
together
groq
transformers>=4.38.0
torch
fastapi
uvicorn[standard]
huggingface-hub
accelerate
open_clip_torch>=2.23
git+https://github.com/openai/CLIP.git
numpy